Revert "Moved save and load methods outside the Main class. Added Serializable to CMDLineUI."
