+++ src/main/java/controllers/GameController.java
@@ -36,25 +36,25 @@ public class GameController implements Serializable, UseCaseOutputBoundary {
         this.order = order;
     }
 
-    public void runPlayerSetUp(List<String> usernames) {
-        for (String s : usernames) {
+    public void runPlayerSetUp(List<String> usernames){
+        for(String s : usernames) {
             boardManager.addPlayer(s);
         }
     }
 
     public void runGame() throws InterruptedException {
         this.ui.printMessage("The playing order is:");
-        for (int i : this.order) {
+        for(int i : this.order) {
             this.ui.printMessage(this.boardManager.getPlayers().get(i).getUsername());
         }
         this.ui.printMessage("Let the game begin! \n");
         TimeUnit.SECONDS.sleep(2);
         // TODO: handle what happens if all players go bankrupt but there are no winners
         // TODO: Remove all instances of boardManager.getPlayers() because cannot access entities in this class
-        while (isWinner() == null | !allBankrupt()) {
+         while (isWinner() == null | !allBankrupt()) {
             for (int i : this.order) {
                 // TODO: Delegate out to helper because this method is large already
-                if (!this.boardManager.getPlayers().get(i).isBankrupt()) {
+                if(!this.boardManager.getPlayers().get(i).isBankrupt()) {
                     // For formatting
                     this.ui.printMessage("\n");
                     this.boardManager.runTurn(i);
@@ -63,27 +63,27 @@ public class GameController implements Serializable, UseCaseOutputBoundary {
                 }
             }
             printCurrentStatistics();
-        }
-        Player winner = isWinner();
-        this.ui.printMessage("Game over! The winner is: ");
-        // TODO: Can't access player here
-        this.ui.printMessage(winner.getUsername());
-        this.ui.printMessage("Thanks for playing!");
+         }
+         Player winner = isWinner();
+         this.ui.printMessage("Game over! The winner is: ");
+         // TODO: Can't access player here
+         this.ui.printMessage(winner.getUsername());
+         this.ui.printMessage("Thanks for playing!");
     }
 
     public Player isWinner() {
-        for (Player p : this.boardManager.getPlayers()) {
-            if (p.getNetWorth() >= this.netWorthGoal) {
+        for(Player p : this.boardManager.getPlayers()) {
+            if(p.getNetWorth() >= this.netWorthGoal) {
                 return p;
             }
         }
         return null;
     }
 
-    public boolean allBankrupt() {
+    public boolean allBankrupt(){
         int num_non_bankrupt = this.order.size();
-        for (Player player : this.boardManager.getPlayers()) {
-            if (player.isBankrupt()) {
+        for(Player player: this.boardManager.getPlayers()){
+            if(player.isBankrupt()){
                 num_non_bankrupt += 1;
             }
         }
@@ -92,7 +92,7 @@ public class GameController implements Serializable, UseCaseOutputBoundary {
 
     public void printCurrentStatistics() throws InterruptedException {
         ui.printMessage("The statistics for this round are: ");
-        for (int i : this.order) {
+        for(int i: this.order) {
             Player player = this.boardManager.getPlayers().get(i);
             int cash = player.getCash();
             int netWorth = player.getNetWorth();
@@ -118,4 +118,7 @@ public class GameController implements Serializable, UseCaseOutputBoundary {
         return this.ui.getInput(messages, acceptedResponses);
     }
 
+    public String getFilepath() {
+        return this.filepath;
+    }
 }
