+++ src/main/java/controllers/GameSetUp.java
@@ -1,8 +1,5 @@
 package controllers;
 
-import details.CmdLineUI;
-
-import java.io.*;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -15,7 +12,7 @@ public class GameSetUp {
      */
     public void setUpGame(UI ui) {
         String input = ui.getStartInfo().toUpperCase();
-        GameController gc = null;
+        GameController gc;
 
         // This branch executes if the user wishes to start a game
         if (input.equals("S")) {
@@ -32,54 +29,17 @@ public class GameSetUp {
         // The only other possible input is Q, which the user types if they wish to quit the game.
         // Nothing needs to happen here if the user wished to quit.
         else {
-            ui.quitGame(null);
+            // This line is included to prevent IntelliJ from giving a warning that the object gc has
+            // not been initialized in the gc.runGame() line below.
+            throw new RuntimeException("This code shouldn't be reached");
         }
         // Run the game
         // runGame method in GameController class throws InterruptedException, so handle that here
-        if (gc != null){
-            try {
-                gc.runGame();
-            } catch(InterruptedException e) {
-                ui.printMessage("Game loop interrupted");
-            }
-        }
-    }
-
-    public static void save(GameController game, String filename){
         try {
-            // opens the file and writes the object into it, then closes the file and prints.
-            FileOutputStream fileOut =
-                    new FileOutputStream(filename + ".txt");
-            ObjectOutputStream out = new ObjectOutputStream(fileOut);
-            out.writeObject(game);
-            out.close();
-            fileOut.close();
-            System.out.println("The game has been saved to " + filename + ".txt in the game directory.");
+            gc.runGame();
+        } catch(InterruptedException e) {
+            ui.printMessage("Game loop interrupted");
         }
-        catch (IOException exception){
-            exception.printStackTrace();
-        }
-    }
-
-    public static GameController load(String filepath){
-        // initializes gc as null which can be returned if the code goes into the catch blocks.
-        GameController gc = null;
-        try {
-            // opens the file and reads the object and allocates it to gc, then closes the file and returns gc.
-            FileInputStream fileIn = new FileInputStream(filepath + ".txt");
-            ObjectInputStream in = new ObjectInputStream(fileIn);
-            gc = (GameController) in.readObject();
-            in.close();
-            fileIn.close();
-
-        } catch (IOException exception) {
-            exception.printStackTrace();
-        }
-        catch (ClassNotFoundException exception){
-            System.out.println("controllers.GameController class not found!");
-            exception.printStackTrace();
-        }
-        return gc;
     }
 
     public ArrayList<Integer> generateOrder(int num_players) {
